+++ CommunitySite/src/main/java/com/portfolio/site/tech/mapper/TechBoardMapper.java
@@ -1,0 +1,9 @@
+package com.portfolio.site.tech.mapper;
+
+import com.portfolio.site.tech.vo.TechBoardVO;
+
+public interface TechBoardMapper {
+	
+	public void write(TechBoardVO params) throws Exception;
+	
+}
