+++ CommunitySite/src/main/java/com/portfolio/site/tech/vo/TechBoardVO.java
@@ -5,10 +5,10 @@ import java.util.Date;
 import lombok.Data;
 
 @Data
-public class TechBoard {
+public class TechBoardVO {
 	private int idx;
 	private String title;
 	private String content;
-	private boolean good;
+	private String writer;
 	private Date regDate;
 }
