+++ CommunitySite/src/main/resources/sqlmapper/board/tech_board-mapper.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.portfolio.site.tech.mapper.TechBoardMapper">
+	<insert id="write" parameterType="TechBoardVO">
+		insert into tb_board_tech (title,content,writer) values (#{title}, #{content}, #{writer});
+	</insert>
+</mapper>
