+++ CommunitySite/src/main/resources/sqlmapper/category/category-mapper.xml
@@ -5,7 +5,7 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 
 <mapper namespace="com.portfolio.site.category.mapper.CategoryMapper">
 
-	<!-- 회원가입 -->
+	<!-- 카테고리 리스트 -->
 	<select id="categoryList" resultType="CategoryVO">
 		SELECT  * FROM tb_category ORDER BY 'order' ASC
 	</select>
